pyproject.toml
[project]
name = "walt-navigator"
version = "1.0.0"
description = "Walt Faqs, Navigation and It support"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "chromadb==1.0.12",
    "deep-translator>=1.11.4",
    "fastapi==0.115.9",
    "langchain>=1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.1",
    "langchain-openai>=1.1.0",
    "langdetect>=1.0.9",
    "langgraph>=1.0.4",
    "numpy==1.26.4",
    "openai>=2.9.0",
    "pydantic>=2",
    "pydantic-settings>=2.12.0",
    "python-dotenv==1.1.0",
    "spacy==3.7.2",
    "unidecode==1.4.0",
    "uvicorn==0.34.3",
]
